+++ lib/src/snackbars/fp_snackbar.dart
@@ -1,169 +1,276 @@
+import 'package:equatable/equatable.dart';
 import 'package:flutter/material.dart';
 import 'package:fp_util/fp_util.dart';
 
-/// snackbar helper class
+class FPSnackBarStyle extends Equatable {
+  final TextStyle textStyle;
+  final TextStyle titleTextStyle;
+  final Color backgroundColor;
+
+  const FPSnackBarStyle({
+    required this.textStyle,
+    required this.titleTextStyle,
+    required this.backgroundColor,
+  });
+
+  @override
+  List<Object?> get props => [
+        textStyle,
+        titleTextStyle,
+        backgroundColor,
+      ];
+}
+
+const _defaultTitleStyle = TextStyle(
+  fontSize: 16,
+  height: 1.25,
+  fontWeight: FontWeight.bold,
+  color: Colors.white,
+);
+const _defaultTextStyle = TextStyle(
+  fontSize: 16,
+  height: 1.25,
+  color: Colors.white,
+);
+
+/// {@template fp_snackbar}
+/// A singleton class to show snackbar
+/// {@endtemplate}
 class FPSnackbar {
-  /// show custom Snackbar
-  static custom(
+  FPSnackbar._() {
+    _defaultStyle = const FPSnackBarStyle(
+      textStyle: _defaultTextStyle,
+      titleTextStyle: _defaultTitleStyle,
+      backgroundColor: Colors.black,
+    );
+
+    /// default success style
+    _successStyle = const FPSnackBarStyle(
+      textStyle: _defaultTextStyle,
+      titleTextStyle: _defaultTitleStyle,
+      backgroundColor: Color(0xFF2D7738),
+    );
+
+    /// default error style
+    _errorStyle = const FPSnackBarStyle(
+      textStyle: _defaultTextStyle,
+      titleTextStyle: _defaultTitleStyle,
+      backgroundColor: Color(0xFF970C0C),
+    );
+
+    /// default info style
+    _infoStyle = const FPSnackBarStyle(
+      textStyle: _defaultTextStyle,
+      titleTextStyle: _defaultTitleStyle,
+      backgroundColor: Color(0xFF136ACA),
+    );
+
+    /// default warning style
+    _warningStyle = const FPSnackBarStyle(
+      textStyle: _defaultTextStyle,
+      titleTextStyle: _defaultTitleStyle,
+      backgroundColor: Color(0xFFF58C20),
+    );
+
+    /// default duration
+    _duration = 3;
+
+    /// default floating
+    _floating = true;
+
+    /// default center message text
+    _centerText = false;
+
+    /// default content max lines
+    _contentMaxLines = 2;
+  }
+
+  static final FPSnackbar _instance = FPSnackbar._();
+
+  factory FPSnackbar() => _instance;
+
+  /// singleton instance
+  static FPSnackbar get instance => _instance;
+
+  /// text style
+  late FPSnackBarStyle _defaultStyle;
+
+  /// set default text style
+  set defaultStyle(FPSnackBarStyle defaultStyle) =>
+      _defaultStyle = defaultStyle;
+
+  /// success style
+  late FPSnackBarStyle _successStyle;
+
+  /// set default title text style
+  set successStyle(FPSnackBarStyle successStyle) =>
+      _successStyle = successStyle;
+
+  /// error style
+  late FPSnackBarStyle _errorStyle;
+
+  /// set default error style
+  set errorStyle(FPSnackBarStyle errorStyle) => _errorStyle = errorStyle;
+
+  /// info style
+  late FPSnackBarStyle _infoStyle;
+
+  /// set default info style
+  set infoStyle(FPSnackBarStyle infoStyle) => _infoStyle = infoStyle;
+
+  /// warning style
+  late FPSnackBarStyle _warningStyle;
+
+  /// set default warning style
+  set warningStyle(FPSnackBarStyle warningStyle) =>
+      _warningStyle = warningStyle;
+
+  /// duration for snackbar dismiss
+  late int _duration;
+
+  /// set default duration
+  set duration(int duration) => _duration = duration;
+
+  /// flag to show floating snackbar
+  late bool _floating;
+
+  /// set default floating
+  set floating(bool floating) => _floating = floating;
+
+  /// message text align
+  late bool _centerText;
+
+  /// set default center message text
+  set centerText(bool centerText) => _centerText = centerText;
+
+  late int? _contentMaxLines;
+
+  /// set default content max lines
+  set contentMaxLines(int? contentMaxLines) =>
+      _contentMaxLines = contentMaxLines;
+
+  /// show default snackbar
+  static void show(
     BuildContext context, {
     required String message,
     String? title,
-    int duration = 3,
-    Color backgroundColor = Colors.black12,
-    TextStyle textStyle = const TextStyle(
-      fontSize: 16,
-      height: 1.25,
-      color: Colors.white,
-    ),
-    TextStyle titleTextStyle = const TextStyle(
-      fontSize: 16,
-      height: 1.25,
-      color: Colors.white,
-      fontWeight: FontWeight.w500,
-    ),
     String? actionText,
     VoidCallback? onActionTap,
-    bool floating = true,
-  }) =>
-      _showSnackbar(
-        context: context,
-        backgroundColor: backgroundColor,
-        textStyle: textStyle,
-        titleTextStyle: titleTextStyle,
-        message: message,
-        title: title,
-        duration: 3,
-        actionText: actionText,
-        onActionTap: onActionTap,
-        floating: floating,
-      );
+  }) {
+    _instance._showSnackbar(
+      context: context,
+      backgroundColor: _instance._defaultStyle.backgroundColor,
+      textStyle: _instance._defaultStyle.textStyle,
+      titleTextStyle: _instance._defaultStyle.titleTextStyle,
+      message: message,
+      title: title,
+      duration: _instance._duration,
+      actionText: actionText,
+      onActionTap: onActionTap,
+      floating: _instance._floating,
+      centerText: _instance._centerText,
+      contentMaxLines: _instance._contentMaxLines,
+    );
+  }
 
   /// show success snackbar
-  static success(
+  static void success(
     BuildContext context, {
     required String message,
     String? title,
     String? actionText,
     VoidCallback? onActionTap,
-    bool floating = true,
-  }) =>
-      _showSnackbar(
-        context: context,
-        backgroundColor: const Color(0xFF2D7738),
-        textStyle: const TextStyle(
-          fontSize: 16,
-          height: 1.25,
-          color: Colors.white,
-        ),
-        titleTextStyle: const TextStyle(
-          fontSize: 16,
-          height: 1.25,
-          fontWeight: FontWeight.bold,
-          color: Colors.white,
-        ),
-        message: message,
-        title: title,
-        duration: 3,
-        actionText: actionText,
-        onActionTap: onActionTap,
-        floating: floating,
-      );
+  }) {
+    _instance._showSnackbar(
+      context: context,
+      backgroundColor: _instance._successStyle.backgroundColor,
+      textStyle: _instance._successStyle.textStyle,
+      titleTextStyle: _instance._successStyle.titleTextStyle,
+      message: message,
+      title: title,
+      duration: _instance._duration,
+      actionText: actionText,
+      onActionTap: onActionTap,
+      floating: _instance._floating,
+      centerText: _instance._centerText,
+      contentMaxLines: _instance._contentMaxLines,
+    );
+  }
 
   /// show error snackbar
-  static error(
+  static void error(
     BuildContext context, {
     required String message,
     String? title,
     String? actionText,
     VoidCallback? onActionTap,
-    bool floating = true,
-  }) =>
-      _showSnackbar(
-        context: context,
-        backgroundColor: const Color(0xFF970C0C),
-        textStyle: const TextStyle(
-          fontSize: 16,
-          height: 1.25,
-          color: Colors.white,
-        ),
-        titleTextStyle: const TextStyle(
-          fontSize: 16,
-          height: 1.25,
-          fontWeight: FontWeight.bold,
-          color: Colors.white,
-        ),
-        message: message,
-        title: title,
-        duration: 3,
-        actionText: actionText,
-        onActionTap: onActionTap,
-        floating: floating,
-      );
-
-  /// show info snackbar
-  static info(
-    BuildContext context, {
-    required String message,
-    String? title,
-    String? actionText,
-    VoidCallback? onActionTap,
-    bool floating = true,
-  }) =>
-      _showSnackbar(
-        context: context,
-        backgroundColor: const Color(0xFF136ACA),
-        textStyle: const TextStyle(
-          fontSize: 16,
-          height: 1.25,
-          color: Colors.white,
-        ),
-        titleTextStyle: const TextStyle(
-          fontSize: 16,
-          height: 1.25,
-          fontWeight: FontWeight.bold,
-          color: Colors.white,
-        ),
-        message: message,
-        title: title,
-        duration: 3,
-        actionText: actionText,
-        onActionTap: onActionTap,
-        floating: floating,
-      );
+  }) {
+    _instance._showSnackbar(
+      context: context,
+      backgroundColor: _instance._errorStyle.backgroundColor,
+      textStyle: _instance._errorStyle.textStyle,
+      titleTextStyle: _instance._errorStyle.titleTextStyle,
+      message: message,
+      title: title,
+      duration: _instance._duration,
+      actionText: actionText,
+      onActionTap: onActionTap,
+      floating: _instance._floating,
+      centerText: _instance._centerText,
+      contentMaxLines: _instance._contentMaxLines,
+    );
+  }
 
   /// show warning snackbar
-  static warning(
+  static void warning(
     BuildContext context, {
     required String message,
     String? title,
     String? actionText,
     VoidCallback? onActionTap,
-    bool floating = true,
-  }) =>
-      _showSnackbar(
-        context: context,
-        backgroundColor: const Color(0xFFF58C20),
-        textStyle: const TextStyle(
-          fontSize: 16,
-          height: 1.25,
-          color: Colors.white,
-        ),
-        titleTextStyle: const TextStyle(
-          fontSize: 16,
-          height: 1.25,
-          fontWeight: FontWeight.bold,
-          color: Colors.white,
-        ),
-        message: message,
-        title: title,
-        duration: 3,
-        actionText: actionText,
-        onActionTap: onActionTap,
-        floating: floating,
-      );
+  }) {
+    _instance._showSnackbar(
+      context: context,
+      backgroundColor: _instance._warningStyle.backgroundColor,
+      textStyle: _instance._warningStyle.textStyle,
+      titleTextStyle: _instance._warningStyle.titleTextStyle,
+      message: message,
+      title: title,
+      duration: _instance._duration,
+      actionText: actionText,
+      onActionTap: onActionTap,
+      floating: _instance._floating,
+      centerText: _instance._centerText,
+      contentMaxLines: _instance._contentMaxLines,
+    );
+  }
+
+  /// show info snackbar
+  static void info(
+    BuildContext context, {
+    required String message,
+    String? title,
+    String? actionText,
+    VoidCallback? onActionTap,
+  }) {
+    _instance._showSnackbar(
+      context: context,
+      backgroundColor: _instance._infoStyle.backgroundColor,
+      textStyle: _instance._infoStyle.textStyle,
+      titleTextStyle: _instance._infoStyle.titleTextStyle,
+      message: message,
+      title: title,
+      duration: _instance._duration,
+      actionText: actionText,
+      onActionTap: onActionTap,
+      floating: _instance._floating,
+      centerText: _instance._centerText,
+      contentMaxLines: _instance._contentMaxLines,
+    );
+  }
 
   /// common method to show remove and show snackbar
-  static void _showSnackbar({
+  void _showSnackbar({
     required BuildContext context,
     required Color backgroundColor,
     required TextStyle textStyle,
@@ -174,6 +281,8 @@ class FPSnackbar {
     required String? actionText,
     required VoidCallback? onActionTap,
     required bool floating,
+    required bool centerText,
+    required int? contentMaxLines,
   }) {
     final effectiveFloating = floating ||
         context.theme.snackBarTheme.behavior == SnackBarBehavior.floating;
@@ -192,6 +301,8 @@ class FPSnackbar {
             actionText: actionText,
             onActionTap: onActionTap,
             floating: effectiveFloating,
+            centerText: centerText,
+            contentMaxLines: contentMaxLines,
           ),
           behavior: effectiveFloating
               ? SnackBarBehavior.floating
@@ -224,6 +335,8 @@ class _SnackbarWidget extends StatelessWidget {
     this.actionText,
     this.onActionTap,
     required this.floating,
+    required this.centerText,
+    this.contentMaxLines,
   });
 
   final Color backgroundColor;
@@ -237,6 +350,10 @@ class _SnackbarWidget extends StatelessWidget {
   final VoidCallback? onActionTap;
 
   final bool floating;
+  final bool centerText;
+  final int? contentMaxLines;
+
+  TextAlign get textAlign => centerText ? TextAlign.center : TextAlign.start;
 
   @override
   Widget build(BuildContext context) {
@@ -258,12 +375,14 @@ class _SnackbarWidget extends StatelessWidget {
         ],
       ),
       child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
+        crossAxisAlignment: CrossAxisAlignment.stretch,
+        mainAxisAlignment: MainAxisAlignment.center,
         children: [
           if (title.isNotNullNotEmpty) ...[
             Text(
               title!,
               style: titleTextStyle,
+              textAlign: textAlign,
             ),
             Sizes.gapV8,
           ],
@@ -274,9 +393,12 @@ class _SnackbarWidget extends StatelessWidget {
               Expanded(
                 child: Text(
                   message,
-                  maxLines: 2,
                   style: textStyle,
-                  overflow: TextOverflow.ellipsis,
+                  maxLines: contentMaxLines,
+                  overflow: contentMaxLines != null
+                      ? TextOverflow.ellipsis
+                      : TextOverflow.visible,
+                  textAlign: textAlign,
                 ),
               ),
               if (actionText != null) ...[